# remotion-composer/ui/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenMontage — Talking-head auto-edit",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/jobs/{job_id}/chat": {
      "get": {
        "summary": "Chat History",
        "operationId": "chat_history_api_jobs__job_id__chat_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Chat History Api Jobs  Job Id  Chat Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Chat Turn",
        "description": "One conversational revise turn. See `_run_chat_turn` for the details.",
        "operationId": "chat_turn_api_jobs__job_id__chat_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs/{job_id}/versions": {
      "get": {
        "summary": "List Versions",
        "operationId": "list_versions_api_jobs__job_id__versions_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Versions Api Jobs  Job Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/cuts": {
      "post": {
        "summary": "Apply Cuts",
        "description": "Deterministic Cắt/Giữ from the transcript editor — no LLM in the loop.\n\nThe editor used to send free text through `/chat`, and whatever the model\ndecided for `nguon` (user vs. director) decided whether the verifier and\nlexicon gate could be bypassed — a model choosing wrong let it silently\nre-block a cut the user explicitly asked for (H3 in the automation-API\nreview). This calls `apply_user_cuts` directly instead: `nguon` is always\n`khach`, unconditionally.",
        "operationId": "apply_cuts_api_jobs__job_id__cuts_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CutsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CutsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/rollback": {
      "post": {
        "summary": "Rollback",
        "description": "Make an earlier version current again, options included.\n\n`versions.rollback_to` copies the spec forward (history stays\nappend-only) and restores every option a later version changed (frame,\nbgm, cut_level, …) — copying `props_vN` too, as this endpoint used to,\nwould leave `src.mp4` (one cut file per job) out of sync with the restored\nspec (H4 in the automation-API review), so a resolve is queued instead.",
        "operationId": "rollback_api_jobs__job_id__rollback_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "turn": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn"
          },
          "applied": {
            "type": "boolean",
            "title": "Applied",
            "default": false
          },
          "queue_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Position"
          },
          "preview_queued": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Queued"
          },
          "remedy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remedy"
          },
          "history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "History"
          },
          "options_changed": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options Changed"
          },
          "not_done": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Not Done"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ChatResponse"
      },
...
      "CutsRequest": {
        "properties": {
          "cut": {
            "items": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Cut"
          },
          "keep": {
            "items": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Keep"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "CutsRequest"
      },
      "CutsResponse": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "queue_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Position"
          },
          "report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Report"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "version"
        ],
        "title": "CutsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "token"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "LoginResponse"
      },
...
      "RollbackRequest": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "version"
        ],
        "title": "RollbackRequest"
      },
      "RollbackResponse": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "reverted_to": {
            "type": "integer",
            "title": "Reverted To"
          },
          "restored_from": {
            "type": "integer",
            "title": "Restored From"
          },
          "options_changed": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options Changed"
          },
          "queue_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Position"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "version",
          "job_id",
          "reverted_to",
          "restored_from"
        ],
        "title": "RollbackResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}